feat: implement algorithm to pick the right audio track for a MediaStreamAudioSourceNode
Previously there was no common behavior when a MediaStream with more than one audio track was used
to create a MediaStreamAudioSourceNode. The Web Audio spec does now provide an algorithm which
defines how to pick an audio track in such a situation.

BREAKING CHANGE: Chrome and Firefox did for example previously just pick the first audio track when
there was more than one. This is fixed now and they behave as defined by the spec.