feat: support ConvolverNodes with a channelCount of 1 or a channelCountMode set to explicit
Previously the ConvolverNode had a fixed channelCount of 2 and a fixed channelCountMode which could
not be changed to anything else but clamped-max. The channelCount can also be 1 and the
channelCountMode can also be explicit.

BREAKING CHANGE: A NotSupportError was previously thrown on any attempt to change the channelCount
or channelCountMode. This is still the case if the channelCount is set to anything above 2 or the
channelCountMode is set to max.