feat: redifne interfaces with TContext
Previously an internal interface called IMinimalBaseAudioContext was used as the minimal required
type whenever an IAudioContext or an IOfflineAudioContext was required. Now TContext is used instead
which is a union of IAudioContext and IOfflineAudioContext.

BREAKING CHANGE: This may break in rare occasions since something which extends
IMinimalBaseAudioContext is not accepted anymore.