fix: call process with an empty array for inactive inputs
Previously the process function of an AudioWorkletProcessor was always called with at least one
channel of audio data for each input even though there was no actively processing AudioNode
connected to this input. This is not the case anymore. The process function is now called with an
empty array if there is no actively processing AudioNode connected to the given input.

BREAKING CHANGE: The process function of an AudioWorkletProcessor might get called with an empty
array now.