fix: enforce PeriodicWave constructor constraints
Previously no checks were applied to the arguments provided to the factory function to create a
PeriodicWave or to its constructor. Now it throws all erros as specified by the spec.

BREAKING CHANGE: In case insufficient arguments were supplied this was not checked before but it
will be checked now.