fix: remove workaround for incorrectly thrown SyntaxError
Safari up to version 13.1 threw a SyntaxError when trying to connect an AudioNode with an AudioNode
of another AudioContext. It also threw a SyntaxError when calling getChannelData() with an
out-of-bounds value and when creating an OfflineAudioContext with invalid values for
numberOfChannels, length, or sampleRate.

BREAKING CHANGE: This will drop support for Safari v13.1.