fix: remove workaround for missing PannerNode property checks
Safari up to version 13.1 allowed to set the channelCount of a PannerNode to a value larger than 2.
It also allowed to set the channelCountMode of a PannerNode to 'max'. It allowed to set the
coneOuterGain of a PannerNode to a value less than 0 or larger than 1. It allowed to set the
maxDistance of a PannerNode to a value less than 0. It allowed to set the refDistance of a
PannerNode to a value less than 0. And it also allowed to set the rolloffFactor of a PannerNode to a
value less than 0.

BREAKING CHANGE: This will drop support for Safari v13.1.