fix: remove workaround to support edge case when calling start()
Safari up to version 13.1 threw an error when calling start() of an AudioBufferSourceNode with an
offset that was equal or greater than the duration of the buffer. It also ignored the offset when
calling start() of an AudioBufferSourceNode with it would have caused the buffer to be not be played
at all.

BREAKING CHANGE: This will drop support for Safari v13.1.